Instead of enter/leave tracking, implement hover(targetHandles)
The problem with enter/leave tracking is that backend has no way to know
if drop targets have changed midflight. For example, drop targets may swap
places, and any retained information about target containment will be outdated.

Now it's up to the backend to specify up-to-date information about drop targets
as it receives this information. We are currently assuming we won't need to do
the same for drop(), but we will see that as we move along with the first backend
implementations.